the node is almost done
i need to finish a few stuff.

I havent tested this yet.
the server is not ready, i need to be working in it.
But for now the most important thing(the node) is almost ready.
Later i will be working in the frontend with a friend because i dont like to work on it and im too lazy to touch css.
If someone is reading this i need to remind you that I love you.
There is no meme I love you.